requirements.txt
arrow==1.2.3
asgiref==3.6.0
binaryornot==0.4.4
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.0.1
click==8.1.3
cookiecutter==2.1.1
crispy-tailwind==0.5.0
cryptography==39.0.0
defusedxml==0.7.1
Django==4.2
django-admin-thumbnails==0.2.6
django-allauth==0.52.0
django-browser-reload==1.6.0
django-ckeditor==6.5.1
django-ckeditor-5==0.2.4
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-embed-video==1.4.8
django-environ==0.9.0
django-filter==22.1
django-js-asset==2.0.0
django-mptt==0.14.0
django-parler==2.3
django-rest-framework==0.1.0
django-rosetta==0.9.9
django-storages==1.13.2
django-tailwind==3.5.0
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
idna==3.4
Jinja2==3.1.2
jinja2-time==0.2.0
MarkupSafe==2.1.2
oauthlib==3.2.2
Pillow==9.4.0
polib==1.2.0
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.6.0
python-dateutil==2.8.2
python-slugify==8.0.0
python3-openid==3.2.0
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.3
stripe==5.0.0
text-unidecode==1.3
urllib3==1.26.14
whitenoise==6.3.0
